Speed-up synchronization process
Remove sync VLAN IDs from sync
Compare vCenter VM list with VNC VM list instead of going one-by-one

Partial-Bug: #1800716
Depends-On: I6188cd450b60ae247a8895add99a3b8fddb54a16

+++ cvm/services.py
@@ -53,11 +53,6 @@ class VirtualMachineInterfaceService(Service):
                                                              vcenter_api_client=vcenter_api_client)
         self._vlan_id_pool = vlan_id_pool
 
-    def sync_vmis(self):
-        self.update_vmis()
-        with self._vcenter_api_client:
-            self._delete_unused_vmis()
-
     def update_vmis(self):
         for vmi_model in list(self._database.vmis_to_update):
             try:
@@ -119,17 +114,6 @@ class VirtualMachineInterfaceService(Service):
 
     def _update_vrouter_port(self, vmi_model):
         self._database.ports_to_update.append(vmi_model)
-
-    def _delete_unused_vmis(self):
-        with self._vcenter_api_client:
-            for vnc_vmi in self._vnc_api_client.get_vmis_by_project(self._project):
-                vmi_model = self._database.get_vmi_model_by_uuid(vnc_vmi.get_uuid())
-                if vmi_model:
-                    continue
-                if self._vcenter_api_client.can_remove_vmi(vnc_vmi):
-                    logger.info('Deleting %s from VNC.', vnc_vmi.name)
-                    self._vnc_api_client.delete_vmi(vnc_vmi.get_uuid())
-                self._delete_vrouter_port(vnc_vmi.get_uuid())
 
     def update_nic(self, nic_info):
         vmi_model = self._database.get_vmi_model_by_uuid(VirtualMachineInterfaceModel.get_uuid(nic_info.macAddress))
@@ -282,15 +266,26 @@ class VirtualMachineService(Service):
                 logger.error('Unexpected exception %s during syncing VM', exc, exc_info=True)
 
     def delete_unused_vms_in_vnc(self):
-        vnc_vms = self._vnc_api_client.get_all_vms()
-        for vnc_vm in vnc_vms:
+        logger.info('Deleting stale information in VNC...')
+        with self._vcenter_api_client:
+            vnc_vm_uuids = self._vnc_api_client.get_all_vm_uuids()
+            vcenter_vms = self._vcenter_api_client.get_all_vms()
+
+            vcenter_vm_uuids = set()
+            for vm in vcenter_vms:
+                try:
+                    vcenter_vm_uuids.add(vm.config.instanceUuid)
+                except Exception, exc:
+                    logger.error('Unexpected exception %s during copying VM info from vCenter.', exc, exc_info=True)
+
+            vms_to_remove = (uuid for uuid in vnc_vm_uuids if uuid not in vcenter_vm_uuids)
+            self._delete_stale_vms_from_vnc(vms_to_remove)
+
+    def _delete_stale_vms_from_vnc(self, vms_to_remove):
+        for uuid in vms_to_remove:
             try:
-                if self._database.get_vm_model_by_uuid(vnc_vm.uuid):
-                    continue
-                with self._vcenter_api_client:
-                    if self._vcenter_api_client.can_remove_vm(uuid=vnc_vm.uuid):
-                        logger.info('Deleting %s from VNC', vnc_vm.name)
-                        self._vnc_api_client.delete_vm(vnc_vm.uuid)
+                logger.info('Deleting stale VM from VNC - uuid: %s', uuid)
+                self._vnc_api_client.delete_vm(uuid=uuid)
             except Exception, exc:
                 logger.error('Unexpected exception %s during removing VM from VNC', exc, exc_info=True)
 
@@ -380,11 +375,11 @@ class VirtualNetworkService(Service):
                 logger.error('Unexpected exception %s during syncing VN', exc, exc_info=True)
 
     def _create_vn_model(self, dpg, vnc_vn):
-            logger.info('Fetched new portgroup key: %s name: %s', dpg.key, vnc_vn.name)
-            vn_model = VirtualNetworkModel(dpg, vnc_vn)
-            self._vcenter_api_client.enable_vlan_override(vn_model.vmware_vn)
-            self._database.save(vn_model)
-            logger.info('Created %s', vn_model)
+        logger.info('Fetched new portgroup key: %s name: %s', dpg.key, vnc_vn.name)
+        vn_model = VirtualNetworkModel(dpg, vnc_vn)
+        self._vcenter_api_client.enable_vlan_override(vn_model.vmware_vn)
+        self._database.save(vn_model)
+        logger.info('Created %s', vn_model)
 
 
 class VRouterPortService(object):
@@ -416,7 +411,9 @@ class VRouterPortService(object):
                 logger.error('Unexpected exception %s during deleting vRouter port', exc, exc_info=True)
 
     def _delete_port(self, uuid):
-        self._vrouter_api_client.delete_port(uuid)
+        port = self._vrouter_api_client.read_port(uuid)
+        if port:
+            self._vrouter_api_client.delete_port(uuid)
 
     def _update_ports(self):
         ports = list(self._database.ports_to_update)
@@ -462,15 +459,6 @@ class VlanIdService(object):
         self._vlan_id_pool = vlan_id_pool
         self._database = database
 
-    def sync_vlan_ids(self):
-        vrouter_uuid = self._esxi_api_client.read_vrouter_uuid()
-        with self._vcenter_api_client:
-            reserved_vlan_ids = self._vcenter_api_client.get_reserved_vlan_ids(vrouter_uuid)
-            for vlan_id in reserved_vlan_ids:
-                self._vlan_id_pool.reserve(vlan_id)
-
-        self.update_vlan_ids()
-
     def update_vlan_ids(self):
         for vmi_model in list(self._database.vlans_to_update):
             try:
@@ -487,7 +475,6 @@ class VlanIdService(object):
                 self._database.vlans_to_restore.remove(vmi_model)
             except Exception, exc:
                 logger.error('Unexpected exception %s during restoring vCenter VLAN', exc, exc_info=True)
-
 
     def _update_vlan_id(self, vmi_model):
         with self._vcenter_api_client:
